added unit testing
