+++ src/components/ComponentForTesting/index.tsx
@@ -1,0 +1,83 @@
+const ComponentForTesting = () => {
+	return (
+		<div
+			className="mb-32 grid text-center lg:max-w-5xl lg:w-full lg:mb-0 lg:grid-cols-4 lg:text-left"
+			data-testid="component-for-testing"
+		>
+			<a
+				href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template-tw&utm_campaign=create-next-app"
+				className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				<h2 className={`mb-3 text-2xl font-semibold`}>
+					Docs{" "}
+					<span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
+						-&gt;
+					</span>
+				</h2>
+				<p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
+					Find in-depth information about Next.js features and API.
+				</p>
+			</a>
+
+			<a
+				href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template-tw&utm_campaign=create-next-app"
+				className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				<h2 className={`mb-3 text-2xl font-semibold`}>
+					Learn{" "}
+					<span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
+						-&gt;
+					</span>
+				</h2>
+				<p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
+					Learn about Next.js in an interactive course
+					with&nbsp;quizzes!
+				</p>
+			</a>
+
+			<a
+				href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template-tw&utm_campaign=create-next-app"
+				className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				<h2 className={`mb-3 text-2xl font-semibold`}>
+					Templates{" "}
+					<span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
+						-&gt;
+					</span>
+				</h2>
+				<p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
+					Discover and deploy boilerplate example
+					Next.js&nbsp;projects.
+				</p>
+			</a>
+
+			<a
+				href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template-tw&utm_campaign=create-next-app"
+				className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				<h2 className={`mb-3 text-2xl font-semibold`}>
+					Deploy{" "}
+					<span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
+						-&gt;
+					</span>
+				</h2>
+				<p
+					className={`m-0 max-w-[30ch] text-sm opacity-50 text-balance`}
+				>
+					Instantly deploy your Next.js site to a shareable URL with
+					Vercel.
+				</p>
+			</a>
+		</div>
+	);
+};
+
+export default ComponentForTesting;
